// src/app/shared/types/api-types/bire-modification-attribute-dto.interface.ts
export interface BireModificationAttributeDTO {
  attributeId?: string;
  attributeValue?: string;
  modificationType?: string;
  modificationNumber?: string;
  modificationRevisionNumber?: string;
  familyCode?: string;
  structureType?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
}