// src/app/shared/types/api-types/h-bire-modification-dto.interface.ts
export interface HBireModificationDTO {
  timestamp?: Date;
  modificationType?: string;
  modificationNumber?: string;
  modificationRevisionNumber?: string;
  modificationTitle?: string;
  familyCode?: string;
  structureType?: string;
  modificationDocumentUrl?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  action?: string;
}